Handle embedded fulltext index in zim file in kiwix.
The checkSearchIndex try to open a SearchIndex on the zim file itself
to see if there is an embedded fulltext.

If there is one (on no already set indexPath in the library), the
library's indexPath is set to the (zim) path.
Then, all "access" to the fulltext index (opening a SearchIndex) will use
the zim's path as indexPath. As the SearchIndex can directly handle
embedded fulltext index, the SearchIndex will open correctly.